use 6 way CAS lock in pin buffer and lwlock ac

+++ src/gausskernel/storage/access/redo/redo_xlogutils.cpp
@@ -1122,7 +1122,7 @@ Block XLogRedoBufferGetPage(RedoBufferManager *buffermanager, Buffer bufferid)
     return blkdata;
 }
 
-void XLogRedoBufferSetState(RedoBufferManager *buffermanager, RedoMemSlot *bufferslot, uint32 state)
+void XLogRedoBufferSetState(RedoBufferManager *buffermanager, RedoMemSlot *bufferslot, uint64 state)
 {
     RedoMemManager *memctl = &(buffermanager->memctl);
     RedoBufferDesc *bufferdesc = NULL;
